% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ibutton.R
\name{read_ibutton_folder}
\alias{read_ibutton_folder}
\title{Read folder of iButton data}
\usage{
read_ibutton_folder(folder, pattern = "*.csv")
}
\description{
Function for reading in a folder of iButton .csv data files, then returning them as a list of data
 @param folder Folder of iButton data to be read in. All files matching the pattern will be included.
 @param pattern All files matching the REGEX pattern will be included. By default, this is any files ending in '.csv' ("*.csv").
}
\details{
@return List of data frames of iButton data, with file name as the identity for each.
 @export
}
